+++ tests/2_ops/prescreen_operation_test.py
@@ -1,0 +1,22 @@
+#!/usr/bin/env python
+"""Tests of the CoLRev prescreen operation"""
+import colrev.review_manager
+
+
+def test_prescreen(  # type: ignore
+    base_repo_review_manager: colrev.review_manager.ReviewManager, helpers
+) -> None:
+    """Test the prescreen operation"""
+
+    helpers.reset_commit(
+        review_manager=base_repo_review_manager, commit="dedupe_commit"
+    )
+
+    prescreen_operation = base_repo_review_manager.get_prescreen_operation()
+    prescreen_operation.create_prescreen_split(create_split=2)
+    prescreen_operation.include_all_in_prescreen(persist=False)
+
+    helpers.reset_commit(
+        review_manager=base_repo_review_manager, commit="dedupe_commit"
+    )
+    prescreen_operation.setup_custom_script()
